requirements.txt
# Multi-Tenant Messaging SaaS - Phase 1
# Production-ready dependencies
# Verified complete dependency list for Python 3.11.9

# Web Framework
fastapi>=0.110.0,<0.120.0  # Upgraded for Pydantic 2.x compatibility
uvicorn[standard]>=0.24.0,<0.30.0
python-multipart>=0.0.6  # Required for FastAPI form data handling (OAuth2PasswordRequestForm)
# Note: starlette and annotated-doc are auto-installed as FastAPI dependencies

# Database
asyncpg>=0.29.0,<0.30.0  # Async PostgreSQL driver
sqlalchemy>=2.0.0,<3.0.0  # SQL ORM
psycopg[binary]>=3.1.0,<4.0.0  # PostgreSQL driver for SQLAlchemy (psycopg3)

# HTTP Client
httpx==0.25.1  # Async HTTP client for Meta OAuth and WhatsApp API

# Configuration & Validation
# Using flexible versions to allow pip to select pre-built wheels
# This prevents Rust compilation errors on Render
pydantic>=2.5.0,<3.0.0  # Data validation
pydantic-settings>=2.1.0  # Settings management
email-validator>=2.0.0  # Required for Pydantic EmailStr validation

# Authentication & Security
python-jose[cryptography]>=3.3.0  # JWT token handling (includes cryptography)
bcrypt>=4.0.0  # Password hashing
pyotp>=2.9.0  # TOTP for 2FA authentication

# AI & LLM
openai>=1.12.0  # OpenAI GPT API client

# Document Processing
python-docx>=1.1.0  # Parse Word documents (.docx)
PyPDF2>=3.0.0  # Parse PDF documents
pdfplumber>=0.10.0  # Better PDF text extraction with tables
openpyxl>=3.1.0  # Parse Excel files (.xlsx)
python-magic-bin>=0.4.14; sys_platform == 'win32'  # File type detection (Windows)
python-magic>=0.4.27; sys_platform != 'win32'  # File type detection (Linux/Mac)
markdown>=3.5.0  # Parse Markdown files

# Rate Limiting & Security
slowapi>=0.1.9  # Rate limiting for FastAPI
redis>=5.0.0  # Optional: For distributed rate limiting (can use in-memory for now)

# Utilities
python-dotenv==1.0.0  # Environment variable management

# Note: All standard library imports (logging, json, os, sys, datetime, typing, etc.)
# are included in Python 3.11.9 and do not need to be listed here.
